Fix the issue Of Status bar
When the status bar is not hidden, the `drawerView` appear after the status bar, and you found a small space between `drawerView` and the top edge of screen. By adding `scrollView.contentInsetAdjustmentBehavior = .never` line in `viewDidLoad()` for iOS 11 and `self.automaticallyAdjustsScrollViewInsets = false` for apps before iOS 11 this issue will be fix.